SCT表修改成绩为浮点型，Course表增加课号查询

+++ server/src/main/java/duan/server/entity/SCT.java
@@ -40,7 +40,7 @@ public class SCT {
 
     private String ctid;
 
-    private Integer grade;
+    private Float grade;
 
     private String fuzzy;
 }
